backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
livekit-agents>=1.0.0
livekit-api>=1.0.0
pydantic>=2.6.0 